separating tests from code
