+++ c/list/IList_01.c
@@ -7,23 +7,6 @@
 #include <string.h>
 #include "IList_01.h"
 
-// typedef int IListValueType;
-
-// typedef struct IListNode
-// {
-//     IListValueType val;
-//     struct IListNode* next;
-// } IListNode;
-
-// typedef struct IList
-// {
-//     IListNode *first;
-//     IListNode *last;
-// } IList;
-
-// IList* IList_new(void);
-// void IList_delete(IList* il);
-// void IList_push_back(IList* il, int val);
 
 IList* IList_new(void)
 {
@@ -71,41 +54,3 @@ void IList_push_back(IList* const il, const int val)
     }
 }
 
-int main(int argc, char* argv[static argc])
-{
-    IList* il = IList_new();
-
-    assert(il != NULL);
-    assert(il->first == NULL);
-    assert(il->last == NULL);
-
-    int x = 11;
-    IList_push_back(il, x);
-    assert(il->first->val == x);
-    assert(il->last->val == x);
-
-    int y = 12;
-    IList_push_back(il, y);
-    assert(il->first->val == x);
-    assert(il->last->next == NULL);
-    assert(il->first->next != NULL);
-
-    assert(il->last->val == y);
-    //printf("val: %d, y: %d\n", il->last->val, y);
-
-    y = 13;
-    IList_push_back(il, y);
-    assert(il->last->val == y);
-    assert(il->last->next == NULL);
-    assert(il->first->next != NULL);
-
-    y = 14;
-    IList_push_back(il, y);
-    assert(il->last->val == y);
-    assert(il->last->next == NULL);
-    assert(il->first->next != NULL);
-
-    IList_delete(il);
-
-    return 0;
-}
